Rename : 모달 컴포넌트 파일 이름 변경
- modal.js -> ProductDetailModal.js로 변경
- 폴더명, 파일명 파스칼케이스로 변경
