+++ src/pages/ProductDetail/index.html
@@ -5,7 +5,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <link rel="stylesheet" href="/src/styles/main.css" />
         <title>상품 상세</title>
-        <script type="module" src="/src/pages/product-detail/index.js"></script>
+        <script type="module" src="/src/pages/ProductDetail/index.js"></script>
     </head>
     <body>
         <c-header></c-header>
