+++ vite.config.js
@@ -13,7 +13,7 @@ export default defineConfig({
                 register: resolve(__dirname, 'src/pages/register/index.html'),
                 login: resolve(__dirname, 'src/pages/login/index.html'),
                 productList: resolve(__dirname, 'src/pages/product-list/index.html'),
-                productDetail: resolve(__dirname, 'src/pages/product-detail/index.html'),
+                productDetail: resolve(__dirname, 'src/pages/ProductDetail/index.html'),
                 cart: resolve(__dirname, 'src/pages/cart/index.html'),
             },
         },
